active admin done

+++ src/app/login/login.component.ts
@@ -1,9 +1,9 @@
-import { Component, OnInit } from '@angular/core';
-import {AuthService} from "../security/auth.service";
-import {TokenService} from "../security/token.service";
-import {Router} from "@angular/router";
-import {SignInForm} from "../security/SignInForm";
-import {FormControl, Validators} from "@angular/forms";
+import {Component, OnInit} from '@angular/core';
+import {AuthService} from '../security/auth.service';
+import {TokenService} from '../security/token.service';
+import {Router} from '@angular/router';
+import {SignInForm} from '../security/SignInForm';
+import {FormControl, Validators} from '@angular/forms';
 
 @Component({
   selector: 'app-login',
@@ -16,7 +16,9 @@ export class LoginComponent implements OnInit {
   form: any = {};
 
   status = '';
-
+  errorLock: any= {
+    message: "LOCK"
+  }
   emailFormControl = new FormControl('', [
     Validators.email, Validators.required
   ]);
@@ -36,36 +38,46 @@ export class LoginComponent implements OnInit {
     this.signInForm = new SignInForm(
       this.form.username,
       this.form.password
-    )
+    );
     this.authService.signIn(this.signInForm).subscribe(data => {
-        if (data.token != undefined) {
+      console.log("dinh" ,data);
+      if(JSON.stringify(data)== JSON.stringify(this.errorLock)) {
+        this.status = 'Tài khoản của bạn đang bị khoá. Chuyển khoản đến stk này để mở khóa. !';
+        return;
+      }
+      if (data.token != undefined) {
         this.tokenService.setIdAccount(data.idAccount);
         this.tokenService.setIdGuest(data.idGuest);
         this.tokenService.setTokenKey(data.token);
         this.tokenService.setNameKey(data.username);
         this.tokenService.setRoleKey(data.roles);
         for (let i = 0; i < this.tokenService.getRoleKey().length; i++) {
-          if (this.tokenService.getRoleKey()[i] == "COMPANY") {
+          if (this.tokenService.getRoleKey()[i] == 'COMPANY') {
             this.router.navigate(['home']).then(() => {
               window.location.reload();
-            })
+            });
           }
-          if (this.tokenService.getRoleKey()[i] == "USER") {
-            this.authService.findById(this.tokenService.getIdAccount()).subscribe(data =>{
-              if(data == "NON_ACTIVE"){
+          if (this.tokenService.getRoleKey()[i] == 'USER') {
+            this.authService.findById(this.tokenService.getIdAccount()).subscribe(data => {
+              if (data == 'NON_ACTIVE') {
                 window.sessionStorage.clear();
-                this.router.navigate(['login']).then(() =>{
+                this.router.navigate(['login']).then(() => {
                   window.location.reload();
-                })
+                });
               }
-            })
+            });
             this.router.navigate(['home']).then(() => {
               window.location.reload();
-            })
+            });
+          }
+          if (this.tokenService.getRoleKey()[i] == 'ADMIN') {
+            this.router.navigate(['list-account']).then(() => {
+              window.location.reload();
+            });
           }
         }
       }
-    })
-    this.status = "Bạn nhập sai tài khoản hoặc mật khẩu!"
+    });
+    this.status = 'Bạn nhập sai tài khoản hoặc mật khẩu!';
   }
 }
